@@ -2,9 +2,9 @@
   import { T } from '@threlte/core';
   import { interactivity } from '@threlte/extras';
   import ObjectItem from './ObjectItem.svelte';
-  import { cameraControls } from '../Camera/stores.ts';
+  let { moveCamera } = $props();
 
-  $: elements = [
+  const elements = [
     {
       name: 'item-0',
       position: { x: -1, y: 1, z: -1 },
@@ -31,26 +31,23 @@
 
   const getMotionPos = (name: string) => {
     const d = elements.filter((element) => element.name === name)[0];
-    return [d.viewPosition.x, d.viewPosition.y, d.viewPosition.z, d.position.x, d.position.y, d.position.z, true];
+    return {
+      viewPosition: d.viewPosition,
+      position: d.position,
+    };
   };
 
   const clickHandler = (e) => {
-    e.stopPropagation();
-    console.log(
-      JSON.stringify({
-        name: e.eventObject.name,
-        position: { ...e.eventObject.parent.position },
-      }),
-    );
+    // console.log(e.eventObject.name);
 
-    // 頻繁に不具合が起きるのでいったんコメントアウト
-    // if (!e.eventObject.isMoved) $cameraControls.setLookAt(...getMotionPos(e.eventObject.name));
-    $cameraControls.setLookAt(...getMotionPos(e.eventObject.name));
+    e.stopPropagation();
+    const payload = getMotionPos(e.eventObject.name);
+    moveCamera(payload);
   };
 </script>
 
 {#each elements as element, index}
   <T.Group key={index} position.x={element.position.x} position.y={element.position.y} position.z={element.position.z}>
-    <ObjectItem name={element.name} on:click={clickHandler} />
+    <ObjectItem name={element.name} onclick={clickHandler} />
   </T.Group>
 {/each}
