@@ -1,61 +1,27 @@
 <script>
   import { T, useTask } from '@threlte/core';
-  import { OrbitControls, Environment, Edges } from '@threlte/extras';
-  import Camera from '../Camera/Camera.svelte';
+  import { Edges } from '@threlte/extras';
   import Frame from '../Frame.svelte';
-  import { baseColor, edgeColor } from './stores.ts';
-  import { cameraControls } from '../Camera/stores.ts';
   import { DoubleSide, MeshNormalMaterial, MeshBasicMaterial } from 'three';
-  import { Pane, Button, Color, Checkbox } from 'svelte-tweakpane-ui';
+  import { GeometorySettings, grid } from './geometory.svelte.ts';
 
-  const grid = 2.8;
-  const initCameraPosition = [0, grid * 0.5, 50];
-  const initCameraLookAt = [0, grid * 0.5, 0];
-
-  let isEdge = true;
-  let isRotation = true;
-  let isNormalMaterial = false;
-  let rotation = 0;
-
+  let rotation = $state(0);
   useTask((delta) => {
-    if (!isRotation) return;
+    if (!GeometorySettings.isRotation) return;
     rotation += delta * 0.5;
   });
 
   const normalMaterial = new MeshNormalMaterial({ side: DoubleSide });
-  $: basicMaterial = new MeshBasicMaterial({ color: $baseColor, side: DoubleSide });
-  $: material = isNormalMaterial ? normalMaterial : basicMaterial;
+  let basicMaterial = $derived(new MeshBasicMaterial({ color: GeometorySettings.baseColor, side: DoubleSide }));
+  let material = $derived(GeometorySettings.isNormalMaterial ? normalMaterial : basicMaterial);
 </script>
-
-<Pane title="Control" position="fixed" y={10}>
-  <Button
-    title="Reset Camera"
-    on:click={() => {
-      if ($cameraControls) $cameraControls.setLookAt(...initCameraPosition, ...initCameraLookAt, true);
-    }}
-  />
-  <Color bind:value={$baseColor} label="Base Color" />
-  <Color bind:value={$edgeColor} label="Edge Color" />
-  <Checkbox bind:value={isNormalMaterial} label="Normal Material" />
-  <Checkbox bind:value={isEdge} label="Edge" />
-  <Checkbox bind:value={isRotation} label="Rotation" />
-</Pane>
-
-<Environment
-  files={'https://raw.githubusercontent.com/mrdoob/three.js/dev/examples/textures/equirectangular/venice_sunset_1k.hdr'}
-  isBackground={false}
-/>
-<T.PerspectiveCamera makeDefault position={[7.5, 2.5, 12.5]} fov={15} far={10000}>
-  <OrbitControls enableZoom={true} enableDamping target.y={1.5} />
-</T.PerspectiveCamera>
-<Camera {initCameraPosition} {initCameraLookAt} />
 
 <T.Group position.x={grid * -2} position.y={0}>
   <T.Group position.x={0} position.y={grid}>
     <Frame label="Box">
       <T.Mesh rotation.y={rotation} {material}>
         <T.BoxGeometry args={[1, 1, 1]} />
-        {#if isEdge}<Edges color={$edgeColor} />{/if}
+        {#if GeometorySettings.isEdge}<Edges color={GeometorySettings.edgeColor} />{/if}
       </T.Mesh>
     </Frame>
   </T.Group>
@@ -63,7 +29,7 @@
     <Frame label="Sphere">
       <T.Mesh rotation.y={rotation} {material}>
         <T.SphereGeometry args={[0.75, 16, 16]} />
-        {#if isEdge}<Edges color={$edgeColor} />{/if}
+        {#if GeometorySettings.isEdge}<Edges color={GeometorySettings.edgeColor} />{/if}
       </T.Mesh>
     </Frame>
   </T.Group>
@@ -71,7 +37,7 @@
     <Frame label="Plane">
       <T.Mesh rotation.y={rotation} {material}>
         <T.PlaneGeometry args={[1, 1]} />
-        {#if isEdge}<Edges color={$edgeColor} />{/if}
+        {#if GeometorySettings.isEdge}<Edges color={GeometorySettings.edgeColor} />{/if}
       </T.Mesh>
     </Frame>
   </T.Group>
@@ -79,7 +45,7 @@
     <Frame label="Circle">
       <T.Mesh rotation.y={rotation} {material}>
         <T.CircleGeometry args={[0.75, 32]} />
-        {#if isEdge}<Edges color={$edgeColor} />{/if}
+        {#if GeometorySettings.isEdge}<Edges color={GeometorySettings.edgeColor} />{/if}
       </T.Mesh>
     </Frame>
   </T.Group>
@@ -87,7 +53,7 @@
     <Frame label="Cone">
       <T.Mesh rotation.y={rotation} {material}>
         <T.ConeGeometry args={[0.5, 1, 16]} />
-        {#if isEdge}<Edges color={$edgeColor} />{/if}
+        {#if GeometorySettings.isEdge}<Edges color={GeometorySettings.edgeColor} />{/if}
       </T.Mesh>
     </Frame>
   </T.Group>
@@ -95,7 +61,7 @@
     <Frame label="Cylinder">
       <T.Mesh rotation.y={rotation} {material}>
         <T.CylinderGeometry args={[0.5, 0.5, 1, 16]} />
-        {#if isEdge}<Edges color={$edgeColor} />{/if}
+        {#if GeometorySettings.isEdge}<Edges color={GeometorySettings.edgeColor} />{/if}
       </T.Mesh>
     </Frame>
   </T.Group>
@@ -103,7 +69,7 @@
     <Frame label="Torus">
       <T.Mesh rotation.y={rotation} {material}>
         <T.TorusGeometry args={[0.5, 0.2, 16, 32]} />
-        {#if isEdge}<Edges color={$edgeColor} />{/if}
+        {#if GeometorySettings.isEdge}<Edges color={GeometorySettings.edgeColor} />{/if}
       </T.Mesh>
     </Frame>
   </T.Group>
@@ -111,7 +77,7 @@
     <Frame label="Ring">
       <T.Mesh rotation.y={rotation} {material}>
         <T.RingGeometry args={[0.5, 0.75, 32]} />
-        {#if isEdge}<Edges color={$edgeColor} />{/if}
+        {#if GeometorySettings.isEdge}<Edges color={GeometorySettings.edgeColor} />{/if}
       </T.Mesh>
     </Frame>
   </T.Group>
@@ -119,7 +85,7 @@
     <Frame label="Icosahedron">
       <T.Mesh rotation.y={rotation} {material}>
         <T.IcosahedronGeometry args={[0.75, 0]} />
-        {#if isEdge}<Edges color={$edgeColor} />{/if}
+        {#if GeometorySettings.isEdge}<Edges color={GeometorySettings.edgeColor} />{/if}
       </T.Mesh>
     </Frame>
   </T.Group>
@@ -127,7 +93,7 @@
     <Frame label="Dodecahedron">
       <T.Mesh rotation.y={rotation} {material}>
         <T.DodecahedronGeometry args={[0.75, 0]} />
-        {#if isEdge}<Edges color={$edgeColor} />{/if}
+        {#if GeometorySettings.isEdge}<Edges color={GeometorySettings.edgeColor} />{/if}
       </T.Mesh>
     </Frame>
   </T.Group>
@@ -135,7 +101,7 @@
     <Frame label="Capsule">
       <T.Mesh rotation.y={rotation} {material}>
         <T.CapsuleGeometry args={[0.375, 0.75, 4, 16]} />
-        {#if isEdge}<Edges color={$edgeColor} />{/if}
+        {#if GeometorySettings.isEdge}<Edges color={GeometorySettings.edgeColor} />{/if}
       </T.Mesh>
     </Frame>
   </T.Group>
@@ -143,7 +109,7 @@
     <Frame label="Tetrahedron">
       <T.Mesh rotation.y={rotation} {material}>
         <T.TetrahedronGeometry args={[0.75, 0]} />
-        {#if isEdge}<Edges color={$edgeColor} />{/if}
+        {#if GeometorySettings.isEdge}<Edges color={GeometorySettings.edgeColor} />{/if}
       </T.Mesh>
     </Frame>
   </T.Group>
