@@ -1,14 +1,30 @@
 <script lang="ts">
-  import ThreeCanvas from '@/components/Three/ThreeCanvas.svelte';
-  import { T } from '@threlte/core';
   import { ContactShadows, Grid, Sky } from '@threlte/extras';
-  import { Pane, Button, Checkbox } from 'svelte-tweakpane-ui';
+  import ThreeCanvas from '@/components/Three/ThreeCanvas.svelte';
   import ObjectGruop from './ObjectGruop.svelte';
-  import Camera from '@/components/Three/Camera/Camera.svelte';
-  import { cameraControls } from '../Camera/stores.ts';
-  import { isTopCamera } from './stores.ts';
-  const initCameraPosition: [number, number, number] = [20, 8.5, 20];
-  const initCameraLookAt: [number, number, number] = [0, 1, 0];
+  import Camera from '../Camera/Camera.svelte';
+  import { Pane, Button } from 'svelte-tweakpane-ui';
+
+  type CameraPayload = {
+    position: {
+      x: number;
+      y: number;
+      z: number;
+    };
+    viewPosition: {
+      x: number;
+      y: number;
+      z: number;
+    };
+  };
+
+  let controls: any;
+  const position = [20, 10, 20];
+  const target = [0, 1, 0];
+  const moveCamera = (payload: CameraPayload) => {
+    controls.setTarget(payload.position.x, payload.position.y, payload.position.z, true);
+    controls.setPosition(payload.viewPosition.x, payload.viewPosition.y, payload.viewPosition.z, true);
+  };
 </script>
 
 <ThreeCanvas>
@@ -16,10 +32,10 @@
     <Button
       title="Reset Camera"
       on:click={() => {
-        $cameraControls.setLookAt(...initCameraPosition, ...initCameraLookAt, true);
+        controls.setTarget(...target, true);
+        controls.setPosition(...position, true);
       }}
     />
-    <Checkbox label="Top Camera" bind:value={$isTopCamera} />
   </Pane>
   <Sky elevation={1} />
   <Grid
@@ -31,17 +47,6 @@
     cellSize={2}
   />
   <ContactShadows scale={10} blur={2} far={2.5} opacity={0.5} />
-  <ObjectGruop />
-
-  {#if $isTopCamera}
-    <T.PerspectiveCamera
-      makeDefault
-      position={[0, 20, 0]}
-      on:create={({ ref }) => {
-        ref.lookAt(0, 1, 0);
-      }}
-    />
-  {:else}
-    <Camera {initCameraPosition} {initCameraLookAt} />
-  {/if}
+  <ObjectGruop {moveCamera} />
+  <Camera bind:controls {position} {target} />
 </ThreeCanvas>
