@@ -1,9 +1,9 @@
 <script>
   import { T } from '@threlte/core';
-  import { objectColor } from './stores';
+  import { MaterialSettings } from './material.svelte.ts';
 </script>
 
 <T.Mesh>
   <T.SphereGeometry />
-  <T.MeshToonMaterial color={$objectColor} />
+  <T.MeshToonMaterial color={MaterialSettings.objectColor} />
 </T.Mesh>
